opt: optimize dtype usage in o2b_deltanet for efficiency and precision
- State (W_t, W_avg) always uses float32 to prevent error accumulation
- Critical computations (matrix inversion, state updates) use float32
- Input tensors (k, q, v, b) keep original dtype when possible
- Simplified dtype handling logic by removing conditional branches
- Updated tests to parametrize both float32 and bfloat16

Precision improvements:
- float32: forward <0.3%, backward <1.5% rel_err
- bfloat16: forward <0.3%, backward <0.4% rel_err

+++ models/o2b_deltanet/layer_o2b_deltanet.py
@@ -74,9 +74,10 @@ class O2DDeltaNetLayer(nn.Module):
         if use_cache:
             cur_state = cache.get('state')
         if cur_state is None:
+            # State uses float32 for numerical stability
             cur_state = (
-                torch.zeros(B, self.n_head, self.head_dim, self.head_dim, device=x.device, dtype=x.dtype),
-                torch.zeros(B, self.n_head, self.head_dim, self.head_dim, device=x.device, dtype=x.dtype),
+                torch.zeros(B, self.n_head, self.head_dim, self.head_dim, device=x.device, dtype=torch.float32),
+                torch.zeros(B, self.n_head, self.head_dim, self.head_dim, device=x.device, dtype=torch.float32),
                 torch.tensor(0, dtype=torch.long, device=x.device),
             )
